migrated common code to e new class, utility functions now use child classes to do their work
