+++ bin/migrate.py
@@ -13,10 +13,8 @@ import time
 import multiprocessing as mp
 import datetime
 
-import filemgmt.compare_utils as compare
+from filemgmt import fmutils
 from filemgmt import migrate_utils as mu
-
-COMPLETE = "Complete"
 
 def parse_cmd_line(argv):
     """ Parse command line arguments
@@ -59,7 +57,6 @@ The following are all valid ways to select the files:
     parser.add_argument('--destination', action='store', required=True, help='Destination folder for the files.')
     parser.add_argument('--current', action='store', default=None, help='Section of the current path that will be replaced by the `destination` path. Examples:\n   --current=\'OPS/\' --destination=\'taiga/\' will move the selected files from <archive root>/OPS to <archive root>/taiga.\n   If left empy the `destinataion` will be prepended to the current directory --destination=\'taiga/\' will move the selected files from <archive root>/ to <archive root>/taiga.\n   --current=\'Y5A1/\' --destination=\'Y5A1_taiga/\' will move <archive root>/OPS/finalcut/Y5A1/XYZ/p02 to <archive root>/OPS/finalcut/Y5A1_taiga/XYZ/p02')
     parser.add_argument('--reqnum', action='store', help='Request number to search for')
-    parser.add_argument('--relpath', action='store', help='Not used')
     parser.add_argument('--unitname', action='store', help='Unit name to search for')
     parser.add_argument('--attnum', action='store', help='Attempt number to search for')
     parser.add_argument('--verbose', action='store_true', help='print differences between db and disk')
@@ -67,10 +64,7 @@ The following are all valid ways to select the files:
     parser.add_argument('--script', action='store_true', help='Print only if there are errors, usefule for running in loops in scripts')
     parser.add_argument('--pfwid', action='store', help='pfw attempt id to search for')
     parser.add_argument('--silent', action='store_true', help='Run with minimal printing, only print ERROR or OK')
-    parser.add_argument('--date_range', action='store', help='Not used')
     parser.add_argument('--tag', action='store', help='Compare all data from a specific tag (this can take a long time)')
-    parser.add_argument('--start_at', action='store', help='Not used', type=int, default=1)
-    parser.add_argument('--end_at', action='store', help='Not used', type=int, default=0)
     parser.add_argument('--dbh', action='store', help=argparse.SUPPRESS) # used internally
     parser.add_argument('--log', action='store', help='Log file to write to, default is to write to sdtout')
     parser.add_argument('--parallel', action='store', help='Specify the parallelization of the migration, e.g. 3 would spread the work across 3 subprocesses.', type=int, default=1)
@@ -79,29 +73,6 @@ The following are all valid ways to select the files:
         cargs.verbose = False
     return cargs
 
-def printProgressBar(win, iteration, count, length = 100, fill = '█', printEnd = "\n"):
-    """ Print a progress bar
-    """
-    percent = (f"{iteration:d}/{count:d}")
-    filledLength = int(length * iteration // count)
-    pbar = fill * filledLength + '-' * (length - filledLength)
-    win.addstr(2, 0, f"Progress: |{pbar}| {percent}{printEnd}")
-
-def run(inputs):
-    """ Method to launch a multiprocessing run
-    """
-    try:
-        (wn, args, pfwids, event, que) = inputs
-        mu.Migration(wn, args, pfwids, event, que)
-    finally:
-        que.put_nowait(mu.Message(wn, COMPLETE, 0))
-
-def results_error(err):
-    """ Error handling routine
-    """
-    print("Exception raised:")
-    print(err)
-    raise err
 
 def main():
     """ Main program module
@@ -109,19 +80,10 @@ def main():
     """
     start = datetime.datetime.now()
     args = parse_cmd_line(sys.argv[1:])
-    if args.date_range:
-        print("Date ranges cannot be used with the migration script")
-        return
-    if args.start_at != 1 or args.end_at != 0:
-        print("start_at and end_at cannot be used with the migration script")
-        return
-    if args.relpath:
-        print("relpath cannot be used with migration script")
-        return
     if args.log is not None:
-        stdp = compare.Print(args.log)
+        stdp = fmutils.Print(args.log)
         sys.stdout = stdp
-    (args, pfwids) = compare.determine_ids(args)
+    (args, pfwids) = fmutils.determine_ids(args)
     manager = mp.Manager()
     event = manager.Event()
 
@@ -147,7 +109,6 @@ def main():
             pos += 1
             if pos >= args.parallel:
                 pos = 0
-        manager = mp.Manager()
         queu = manager.Queue()
         done = [False] * len(jobs)
         wins = []
@@ -161,7 +122,7 @@ def main():
                 wins.append(curses.newwin(step, num_cols, i*step, 0))
 
             with mp.Pool(processes=len(jobs), maxtasksperchild=1) as pool:
-                _ = [pool.apply_async(run, args=((i, copy.deepcopy(args), jobs[i], event, queu,),), error_callback=results_error) for i in range(len(jobs))]
+                _ = [pool.apply_async(fmutils.run, args=((mu.Migration, i, copy.deepcopy(args), jobs[i], event, queu,),), error_callback=fmutils.results_error) for i in range(len(jobs))]
                 pool.close()
                 #pool.join()
                 while not all(done):
@@ -173,7 +134,7 @@ def main():
                                     errors[ms.pfwid] = []
                                 errors[ms.pfwid].append(ms.msg)
                                 continue
-                            if ms.msg == COMPLETE:
+                            if ms.msg == fmutils.COMPLETE:
                                 done[ms.win] = True
                                 wins[ms.win].clear()
                                 wins[ms.win].addstr("Complete\n")
@@ -181,7 +142,7 @@ def main():
                                 wins[ms.win].clear()
                                 wins[ms.win].addstr(ms.msg + '\n')
                             else:
-                                printProgressBar(wins[ms.win], ms.iteration, ms.count)
+                                fmutils.printProgressBar(wins[ms.win], ms.iteration, ms.count)
                             wins[ms.win].refresh()
                         except queue.Empty:
                             break
