+++ python/filemgmt/disk_utils_local.py
@@ -158,142 +158,6 @@ def remove_file_if_exists(filename):
 # end remove_file_if_exists
 
 ######################################################################
-def get_files_from_disk(relpath, archive_root, check_md5sum=False, debug=False):
-    """ Check disk to get list of files within that path inside the archive
-
-        Parameters
-        ----------
-        archive_root : str
-            The base root of the relpath entry
-
-        check_md5sum : bool
-            Whether or not to compare md5sums
-
-        debug : bool
-            Whether or not to report debugging info
-
-        Returns
-        -------
-        A dictionary contianing the info about the files on disk (filesize, md5sum, compression, filename, path)
-
-    """
-
-    start_time = time.time()
-    if debug:
-        print("Getting file information from disk: BEG")
-
-    files_from_disk = {}
-    duplicates = {}
-    for (dirpath, _, filenames) in os.walk(os.path.join(archive_root, relpath)):
-        for filename in filenames:
-            fullname = f'{dirpath}/{filename}'
-            data = get_single_file_disk_info(fullname, check_md5sum, archive_root)
-            if filename in files_from_disk:
-                if filename not in duplicates:
-                    duplicates[filename] = [copy.deepcopy(files_from_disk[filename])]
-                duplicates[filename].append(data)
-                #print "DUP",filename,files_from_disk[filename]['path'],data['path']
-            else:
-                files_from_disk[filename] = data
-
-    end_time = time.time()
-    if debug:
-        print(f"Getting file information from disk: END ({end_time - start_time} secs)")
-    return files_from_disk, duplicates
-
-####################################################################
-def compare_db_disk(files_from_db, files_from_disk, duplicates, check_md5sum, debug=False, archive_root="",
-                    pfwid=None):
-    """ Compare file info from DB to info from disk
-
-        Parameters
-        ----------
-        file_from_db : dict
-            Dicitonary containing the file info from the database
-
-        files_from_disk : dict
-            Dictionary containing the file info from disk
-
-        check_md5sum : bool
-            Whether or not to report the md5sum comparision
-
-        check_filesize : bool
-            Whether or not to report the filesize comparison
-
-        debug : bool
-            Whenther or not to report debugging info
-            Default: False
-
-        archive_root : str
-            The archive root path
-            Default : False
-
-        Returns
-        -------
-        None
-    """
-
-    start_time = time.time()
-    if debug:
-        print("Comparing file information: BEG")
-    comparison_info = {'equal': [],
-                       'dbonly': [],
-                       'diskonly': [],
-                       'both': [],
-                       'path': [],
-                       'filesize': [],
-                       'duplicates': [], # has db entry
-                       'pathdup' : [],    # has no db entry
-                       'pfwid' : pfwid
-                      }
-    if check_md5sum:
-        comparison_info['md5sum'] = []
-
-    #if check_filesize:
-    #    comparison_info['filesize'] = []
-
-    allfiles = set(files_from_db).union(set(files_from_disk))
-
-    for fname in allfiles:
-        if fname in files_from_db:
-            if fname in files_from_disk:
-                comparison_info['both'].append(fname)
-                fdisk = files_from_disk[fname]
-                fdb = files_from_db[fname]
-                if fname in duplicates:
-                    comparison_info['duplicates'].append(fname)
-                if fdisk['relpath'] == fdb['path']:
-                    if fdisk['filesize'] == fdb['filesize']:
-                        if check_md5sum:
-                            if fdisk['md5sum'] == fdb['md5sum']:
-                                comparison_info['equal'].append(fname)
-                            else:
-                                comparison_info['md5sum'].append(fname)
-                        else:
-                            comparison_info['equal'].append(fname)
-                    else:
-                        comparison_info['filesize'].append(fname)
-                else:
-                    try:
-                        data = get_single_file_disk_info(fdb['path'] + '/' + fname, check_md5sum, archive_root)
-                        if fname not in duplicates:
-                            duplicates[fname] = []
-                        duplicates[fname].append(copy.deepcopy(files_from_disk[fname]))
-                        files_from_disk[fname] = data
-                        comparison_info['duplicates'].append(fname)
-                    except:
-                        comparison_info['path'].append(fname)
-            else:
-                comparison_info['dbonly'].append(fname)
-        else:
-            if fname in duplicates:
-                comparison_info['pathdup'].append(fname)
-            comparison_info['diskonly'].append(fname)
-
-    end_time = time.time()
-    if debug:
-        print(f"Comparing file information: END ({end_time - start_time} secs)")
-    return comparison_info
 
 def del_files_from_disk(path):
     """ Delete files from disk """
